+++ CHANGELOG.md
@@ -12,8 +12,9 @@ distances;
 - Function `trmchangeangles` for convenient angle change in a model;
 - Option to leave only alpha carbon atoms added to `pdbbackbone`;
 - Option to consider planar angles added to `trmdistantangleindices`;
-- `trmcreate` creates a alpha carbon-based transformation model if the
+- `trmcreate` creates an alpha carbon-based transformation model if the
 specified PDB structures contains only alpha carbon atoms;
+- `pdbcost` supports PDB files containing alpha carbon-based structures;
 - `trmplottranglediff` replaced with the `trmplotanglediff` function that
 supports both planar and torsion angles.